Split invokeai/backend/image_util/segment_anything/ dir into grounding_dino/ and segment_anything/
