+++ invokeai/backend/image_util/grounding_dino/detection_result.py
